Fixes and improvements to ChoiceLetter_WildMenJoin
Preventing right click dismiss will force the player to make a choice. Currently the player can dismiss the letter without any repercussions, avoiding either taking extra pawns or risking the raid chance. They'd still be able to return to the letter at any time in the future as well. An additional change that could be done here (but I haven't done it) would be to set a timeout as well which would force the player to make a choice in a timely manner.

Using the `map` field instead of `Find.CurrentMap` was done mostly for consistency. The rejection raid will hit a specific player settlement, while accepting the pawns can be done at any map whatsoever (including temporary quest maps).

Exposing data (namely the list of pawns and the map) should fix the issues caused by saving and reloading the game. Without those changes, saving and reloading would cause the player to be unable to accept the pawns (as they were not saved), and rejecting had a chance to cause an error as well (in case a raid was chosen), but that option could be repeatedly picked until successfully declining.