Small improvements to large fire
First change is to only increment the large fire counter for player controlled fires. This will allow those fires to be used in custom non-player settlements and quests without interfering with player faction - mostly for compatibility with other mods that decide to use this building.

The second change is to spawn the wild pawn for a random map with a player controlled large fire, giving each map an equal chance to get the event. The current implementation will only ever spawn wild people on one map over and over (unless the fire was put out) due to everything always ticking in the same, deterministic order.