Added inactive main map/no active maps fallback for ChoiceLetter_WildMenJoin
The ChoiceLetter will first check if the target map is null or not in the list of active maps. It'll then pick a random player home map (ignoring non-player home maps to avoid potential exploits, like spawning raids when rejecting on quest maps).

The accept option will only be active if there's any player map active. The reject option will always be available, but raids will only be possible if there's any player map.

The changes done here are to prevent issues when a player abandons the map the quest/letter was tied to, as well as an additional one if they abandon all their maps.